fixed problem with db

+++ web/db/migrate/20170920191707_create_trackers.rb
@@ -1,8 +1,8 @@
 class CreateTrackers < ActiveRecord::Migration[5.1]
   def change
     create_table :trackers do |t|
-      t.integer :lat
-      t.integer :lon
+      t.float :lat
+      t.float :lon
 
       t.timestamps
     end
